chore : 네이밍 수정
