+++ src/main/java/cart/common/argumentresolver/MemberBody.java
@@ -1,0 +1,12 @@
+package cart.common.argumentresolver;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface MemberBody {
+
+}
